# use `jl_gf_invoke_lookup` in `invoke_tfunc` to lookup the method to be called (`invoke` improvement No. 2)