This is the second one from the break down of my old PR (#9642) to improve the performance of `invoke`. The original commit was [here](https://github.com/yuyichao/julia/commit/b0d5045a7e703de916398a49dcae9a0fb92fc4f8).

This particular commit might not have much impact on the performance but IMHO, since we already have `jl_gf_invoke_lookup` to find the method to be called by invoke, we might as well use it to simplify the code.