@yuyichao 2015-05-17
Ping. Can anyone review this please?

For performance, this one should be much faster.
For correctness, the new one should by definition be more correct since it is using the same function the real `invoke` uses for method lookup.